Start making changes to make this a standalone repo
This changes the top level of the repository to make it behave more like a standalone repo,
while still being compatible with the auto updates which come in from the pipeline to the main source code.

Gets travis working, tests, and some basic documentation. There will likely be more that comes later, but this is some of the core requirements we will need to tackle eventually anyways.

Also introduces some changes to the PR template, adding a check mark that PR's do not try to change the source code.

Depends on MolSSI/QCFractal#56 to get the test imports working correctly.