Start process of converting to Pydantic for validation
This commit starts the pydantic validation conversion by adding
the BaseModel from Pydantic to the Collections, and reworking
the collections to not rely on an abstract method + extra hidden
properties to infer data model and validation.

This is not the complete implementation, but should get it started.
Outstanding TODOs (can be completed in separate PRs, this is stable
now):
* Convert the calls from interface <-> client to handle JSON string instead of dict
* Ensure the validation of Collection DataModels is correctly specific, especially on Dict models
* Double check the OpenFFWorkflow data model for correct defaults
* Add Pydantic validation to Procedures