Better doc-formatter tech and model auto-doc's
Refactored the doc-formatter to be more general, even though it wound
up being a bunch of if...elif statements.

This also adds the `models.rst` file which just tries to run anything
from a BaseModel class through the doc formatter.

The formatter is now wrapped around a try...except so if upstream
changes break the rather fragile flow, then it won't bring the
functioning tech down.

I also added a test for the auto-doc string to make sure everything
at least functions.