Replace curl with wget in travis
Curl seems to be having a problem with some redirects on
continuum.io. Changing the URL to anaconda.com seems to
work, at least for now.

This affects the MD5 test of the downloaded miniconda environment,
causing all travis builds to fail.

I also change the curl command to wget to make it consistent with
how the file itself is retrieved.